redpillx/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redpillx"
version = "0.1.3"
description = "AI-powered generic SDK for dynamic chart generation from any JSON data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Shanthosh", email = "shanthubolt@gmail.com"}
]

dependencies = [
    "polars>=0.20.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "langchain-core>=0.3.0",
]

[project.optional-dependencies]
langgraph = [
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-aws>=0.1.0",
    "langchain-ollama>=0.1.0",
    "langgraph>=0.2.0",
]

all = [
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-aws>=0.1.0",
    "langchain-ollama>=0.1.0",
    "langgraph>=0.2.0",
    "google-generativeai>=0.8.0",
    "cohere>=5.0.0",
]

dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "build>=1.0.0",
    "twine>=5.0.0",
    "wheel>=0.42.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true